+++ SwiftyRSA/NSData+SHA1.m
@@ -1,0 +1,22 @@
+//
+//  NSData_SHA1.h
+//  SwiftyRSA
+//
+//  Created by Paul Wilkinson on 19/04/2016.
+//  Copyright © 2016 Scoop. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+#import <CommonCrypto/CommonCrypto.h>
+
+@implementation NSData (NSData_SHA1)
+
+- (nonnull NSData*) SHA1 {
+    unsigned int outputLength = CC_SHA1_DIGEST_LENGTH;
+    unsigned char output[outputLength];
+    
+    CC_SHA1(self.bytes, (unsigned int) self.length, output);
+    return [NSData dataWithBytes:output length:outputLength];
+}
+
+@end