+++ SwiftyRSATests/SwiftyRSAObjcTests.m
@@ -104,4 +104,70 @@
     XCTAssertTrue([data isEqualToData:decrypted]);
 }
 
+- (void)testSignVerify {
+    
+    NSMutableData* data = [NSMutableData dataWithCapacity:2048 * sizeof(UInt32)];
+    for (unsigned int i = 0 ; i < 2048 ; ++i ){
+        u_int32_t randomBits = arc4random();
+        [data appendBytes:(void*)&randomBits length:sizeof(UInt32)];
+    }
+    
+    NSString* pubString = [TestUtils pemKeyStringWithName:@"swiftyrsa-public"];
+    NSString* privString = [TestUtils pemKeyStringWithName:@"swiftyrsa-private"];
+    NSData* pubData = [TestUtils derKeyDataWithName:@"swiftyrsa-public"];
+    
+    NSError* error;
+    
+    NSString* signature = [SwiftyRSA signString:[data description] privateKeyPEM:privString error:&error];
+    
+    XCTAssertNil(error);
+
+    VerificationResult* result = [SwiftyRSA verifySignatureString:[data description] signature:signature publicKeyPEM:pubString error:&error];
+    
+    XCTAssertNil(error);
+    XCTAssert(result.boolValue);
+    
+    result = [SwiftyRSA verifySignatureString:[data description] signature:signature publicKeyDER:pubData error:&error];
+    
+    XCTAssertNil(error);
+    XCTAssert(result.boolValue);
+    
+    NSData *signatureData = [SwiftyRSA signData:data privateKeyPEM:privString error:&error];
+    
+    XCTAssertNil(error);
+    
+    result = [SwiftyRSA verifySignatureData:data signature:signatureData publicKeyPEM:pubString error:&error];
+    
+    XCTAssertNil(error);
+    XCTAssert(result.boolValue);
+    
+    result = [SwiftyRSA verifySignatureData:data signature:signatureData publicKeyDER:pubData error:&error];
+    
+    XCTAssertNil(error);
+    XCTAssert(result.boolValue);
+    
+    SwiftyRSA* rsa=[[SwiftyRSA alloc]init];
+    
+    SecKeyRef pubKey = [rsa publicKeyFromPEMString:pubString error:&error];
+    
+    XCTAssertNil(error);
+    
+    SecKeyRef privKey = [rsa privateKeyFromPEMString:privString error:&error];
+    
+    XCTAssertNil(error);
+    
+    NSData* digest=[data SHA1];
+    
+    NSData* digestSignature = [rsa signSHA1Digest:digest privateKey:privKey error:&error];
+    
+    XCTAssertNil(error);
+    
+    result = [rsa verifySHA1SignatureData:digest signature:digestSignature publicKey:pubKey error:&error];
+    
+    XCTAssertNil(error);
+    XCTAssert(result.boolValue);
+
+
+   }
+
 @end