+++ SwiftyRSATests/SwiftyRSATests.swift
@@ -106,4 +106,55 @@ class SwiftyRSATests: XCTestCase {
         
         XCTAssertEqual(data, decrypted)
     }
+    
+    func testSignVerify() {
+        
+        
+        let bytes = [UInt32](count: 2048, repeatedValue: 0).map { _ in arc4random() }
+        let data = NSData(bytes: bytes, length: bytes.count * sizeof(UInt32))
+        
+        let pubString = TestUtils.pemKeyString(name: "swiftyrsa-public")
+        let privString = TestUtils.pemKeyString(name: "swiftyrsa-private")
+        
+        let pubData = TestUtils.derKeyData(name: "swiftyrsa-public")
+        
+        let signature = try! SwiftyRSA.signData(data, privateKeyPEM: privString)
+        
+        var result = try! SwiftyRSA.verifySignatureData(data, signature: signature, publicKeyPEM: pubString)
+        XCTAssert(result)
+        
+        result = try! SwiftyRSA.verifySignatureData(data, signature:  signature, publicKeyDER:  pubData)
+        XCTAssert(result)
+        
+        let badBytes = [UInt32](count: 16, repeatedValue: 0).map { _ in arc4random() }
+        let badData = NSData(bytes: badBytes, length: badBytes.count * sizeof(UInt32))
+        
+        result = try! SwiftyRSA.verifySignatureData(badData, signature:  signature, publicKeyPEM: pubString)
+        XCTAssert(!result)
+        
+        let testString = "Lorum Ipsum Ipso Facto Ad Astra Ixnay Onay Ayway"
+        
+        let stringSignature = try! SwiftyRSA.signString(testString, privateKeyPEM: privString)
+        
+        result = try! SwiftyRSA.verifySignatureString(testString, signature: stringSignature, publicKeyPEM: pubString)
+        XCTAssert(result)
+        
+        result = try! SwiftyRSA.verifySignatureString(testString, signature: stringSignature, publicKeyDER: pubData)
+        XCTAssert(result)
+        
+        let rsa = SwiftyRSA()
+        
+        let pubKey = try! rsa.publicKeyFromPEMString(pubString)
+        let privKey = try! rsa.privateKeyFromPEMString(privString)
+        
+        let digest=data.SHA1()
+        
+        let digestSignature = try! rsa.signSHA1Digest(digest, privateKey: privKey)
+        
+        result = try! rsa.verifySHA1SignatureData(digest, signature: digestSignature, publicKey: pubKey)
+        XCTAssert(result)
+        
+        
+
+    }
 }